+++ traits_futures/qt/tests/test_pingee.py
@@ -22,9 +22,9 @@ from traits_futures.tests.i_pingee_tests import IPingeeTests
 @requires_qt
 class TestPingee(GuiTestAssistant, IPingeeTests, unittest.TestCase):
     def toolkit_factory(self):
-        from traits_futures.qt.init import QtToolkit
+        from traits_futures.qt.context import QtContext
 
-        return QtToolkit()
+        return QtContext()
 
     def setUp(self):
         GuiTestAssistant.setUp(self)
