+++ traits_futures/wx/context.py
@@ -9,21 +9,26 @@
 # Thanks for using Enthought open source!
 
 """
-Interface for GUI toolkit objects.
+Entry point for finding toolkit-specific classes.
 """
+# Force an ImportError if wxPython is not installed.
+import wx  # noqa: F401
 
-import abc
+from traits_futures.i_gui_context import IGuiContext
+from traits_futures.wx.event_loop_helper import EventLoopHelper
+from traits_futures.wx.pingee import Pingee
 
 
-class IToolkit(abc.ABC):
-    @abc.abstractmethod
+@IGuiContext.register
+class WxContext:
     def pingee(self, on_ping):
         """
         Return a new Pingee instance for this toolkit.
         """
+        return Pingee(on_ping=on_ping)
 
-    @abc.abstractmethod
     def event_loop_helper(self):
         """
         Return a new EventLoopHelper instance for this toolkit.
         """
+        return EventLoopHelper()
