+++ traits_futures/wx/tests/test_pingee.py
@@ -22,9 +22,9 @@ from traits_futures.tests.i_pingee_tests import IPingeeTests
 @requires_wx
 class TestPingee(GuiTestAssistant, IPingeeTests, unittest.TestCase):
     def toolkit_factory(self):
-        from traits_futures.wx.init import WxToolkit
+        from traits_futures.wx.context import WxContext
 
-        return WxToolkit()
+        return WxContext()
 
     def setUp(self):
         GuiTestAssistant.setUp(self)
